formatting and code review fixes
