+++ lib/powerpoint.rb
@@ -1,10 +1,13 @@
 require 'powerpoint/version'
 require 'powerpoint/util'
 require 'powerpoint/slide/intro'
-require 'powerpoint/slide/content'
+require 'powerpoint/slide/text_content'
+require 'powerpoint/slide/default_slide'
+require 'powerpoint/slide/image_content'
+require 'powerpoint/slide/table_content'
+require 'powerpoint/slide/text'
 require 'powerpoint/compression'
 require 'powerpoint/presentation'
-require 'pry'
 
 module Powerpoint
   ROOT_PATH = File.expand_path("../..", __FILE__)
