+++ lib/powerpoint/slide/table_content.rb
@@ -1,0 +1,22 @@
+require 'erb'
+
+module Powerpoint
+  module Slide
+    class TableContent
+      include Powerpoint::Util
+
+      attr_reader :content, :id, :idx
+      attr_writer :id, :idx
+
+      def initialize(options={})
+        require_arguments [:content, :idx], options
+        options.each { |k, v| instance_variable_set("@#{k}", v) }
+
+      end
+
+      def render
+        render_str('table.xml.erb')
+      end
+    end
+  end
+end
