+++ lib/powerpoint/views/picture.xml.erb
@@ -1,40 +1,40 @@
 <p:pic>
-<p:nvPicPr>
-  <p:cNvPr id="<%= id %>" name="Content Placeholder 3"/>
-  <p:cNvPicPr>
-    <a:picLocks noChangeAspect="1" noGrp="1"/>
-  </p:cNvPicPr>
-  <% if idx > 0 %>
-  <p:nvPr>
-    <p:ph idx="<%= idx %>"/>
-  </p:nvPr>
+  <p:nvPicPr>
+    <p:cNvPr id="<%= id %>" name="Content Placeholder 3"/>
+    <p:cNvPicPr>
+      <a:picLocks noChangeAspect="1" noGrp="1"/>
+    </p:cNvPicPr>
+    <% if idx > 0 %>
+      <p:nvPr>
+        <p:ph idx="<%= idx %>"/>
+      </p:nvPr>
+    <% else %>
+      <p:nvPr/>
+    <% end %>
+  </p:nvPicPr>
+  <p:blipFill>
+    <a:blip r:embed="<%= embed_rid %>">
+      <a:extLst>
+        <a:ext uri="{28A0092B-C50C-407E-A947-70E740481C1C}">
+          <a14:useLocalDpi val="0" xmlns:a14="http://schemas.microsoft.com/office/drawing/2010/main"/>
+        </a:ext>
+      </a:extLst>
+    </a:blip>
+    <a:stretch>
+      <a:fillRect/>
+    </a:stretch>
+  </p:blipFill>
+  <% if coords.any? %>
+    <p:spPr>
+      <a:xfrm>
+        <a:off x="<%= coords[:x].to_s %>" y="<%= coords[:y].to_s %>"/>
+        <a:ext cx="<%= coords[:cx].to_s %>" cy="<%= coords[:cy].to_s %>"/>
+      </a:xfrm>
+      <a:prstGeom prst="rect">
+        <a:avLst/>
+      </a:prstGeom>
+    </p:spPr>
   <% else %>
-  <p:nvPr/>
+    <p:spPr/>
   <% end %>
-</p:nvPicPr>
-<p:blipFill>
-  <a:blip r:embed="<%= embed_rid %>">
-    <a:extLst>
-      <a:ext uri="{28A0092B-C50C-407E-A947-70E740481C1C}">
-        <a14:useLocalDpi val="0" xmlns:a14="http://schemas.microsoft.com/office/drawing/2010/main"/>
-      </a:ext>
-    </a:extLst>
-  </a:blip>
-  <a:stretch>
-    <a:fillRect/>
-  </a:stretch>
-</p:blipFill>
-<% if coords.any? %>
-<p:spPr>
-  <a:xfrm>
-    <a:off x="<%= coords[:x].to_s %>" y="<%= coords[:y].to_s %>"/>
-    <a:ext cx="<%= coords[:cx].to_s %>" cy="<%= coords[:cy].to_s %>"/>
-  </a:xfrm>
-  <a:prstGeom prst="rect">
-    <a:avLst/>
-  </a:prstGeom>
-</p:spPr>
-<% else %>
-<p:spPr/>
-<% end %>
 </p:pic>
