+++ lib/powerpoint/views/rels.xml.erb
@@ -2,6 +2,6 @@
 <Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">
   <Relationship Id="rId1" Target="../slideLayouts/slideLayout<%=slide_layout%>.xml" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/slideLayout"/>
   <% images.each do |image| %>
-  <Relationship Id="<%= image.embed_rid %>" Target="../media/<%= image.image_name %>" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/image"/>
+    <Relationship Id="<%= image.embed_rid %>" Target="../media/<%= image.image_name %>" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/image"/>
   <% end %>
 </Relationships>
