+++ lib/powerpoint/views/text.xml.erb
@@ -1,28 +1,29 @@
 <p:sp>
-<p:nvSpPr>
-  <p:cNvPr id="<%= id %>" name="Content Placeholder 2"/>
-  <p:cNvSpPr>
-    <a:spLocks noGrp="1"/>
-  </p:cNvSpPr>
-  <p:nvPr>
-    <p:ph idx="<%= idx %>"/>
-  </p:nvPr>
-</p:nvSpPr>
-<p:spPr/>
-<p:txBody>
-  <% unless bullets %>
-  <a:bodyPr>
-    <a:normAutofit/>
-  </a:bodyPr>
-  <% else %>
-  <a:bodyPr/>
-  <% end %>
-  <a:lstStyle/>
-  <% content.each do |content| %>
-  <%= content.render(bullets) %>
-  <!--<a:endParaRPr lang="en-US"/>-->
-  <% end %>
+  <p:nvSpPr>
+    <p:cNvPr id="<%= id %>" name="Content Placeholder 2"/>
+    <p:cNvSpPr>
+      <a:spLocks noGrp="1"/>
+    </p:cNvSpPr>
+    <p:nvPr>
+      <p:ph idx="<%= idx %>"/>
+    </p:nvPr>
+  </p:nvSpPr>
+  <p:spPr/>
+  <p:txBody>
 
+    <% unless bullets %>
+      <a:bodyPr>
+        <a:normAutofit/>
+      </a:bodyPr>
+    <% else %>
+      <a:bodyPr/>
+    <% end %>
 
-</p:txBody>
+    <a:lstStyle/>
+
+    <% content.each do |content| %>
+      <%= content.render(bullets) %>
+    <% end %>
+
+  </p:txBody>
 </p:sp>
